Add Django-like models module
This will be used only for API event classes. The use of models
in structures resuted in messy and complicated code in the past.

Using __init__ is uglier but simpler, and there are no hidden
mechanics happening in the background which require a higher time
of understanding, using these for API events only decreases the
effort.

These modules will me moved to a different folder soon.