Refactor Model into an EventAdapter
Model was meant to be used with structures at first.
However, code complexity would increase a lot. For example,
I have a message object which contains the ID of the room, why
would we create some ugly code which uses that Field and converts
it into a Room object instead of just using a simple
state.get_room_by_id() function.