Port ban message RegEx to free-spacing mode
RegEx is hard to read - thus I am improving readability by adding
the VERBOSE, also known as X, flag so we can add comments to it and
make spaces explicit.
This RegEx shouldn't even exist, but Chatovod doesn't return
the ban list as a JSON object so we've got to parse the text
to extract the ban information.